Address PR #11 review comments - Move METS_BROWSING/METS2_BROWSING XML files into tekst object folder - Remove primary/access fallback paths in TestObjectFolderHelper - Replace symlinks with Files.copy in TestObjectFolderHelper - Fix XSD validation failure not affecting routing (set isValid=false) - Remove 'recursively' from CapabilityDescription (Files.list is non-recursive) - Add METAe/xlink.xsd mapping in XsdValidator for offline ALTO validation - Update all test file references to new paths
