+++ nifi-tekst-bundle-processors/src/test/kotlin/no/nb/nifi/tekst/processors/JhoveTest.kt
@@ -21,7 +21,7 @@ class JhoveTest {
     private val fixtureRoot = Paths.get(projectFolder, "src/test/resources/tekst_ee11f8dd-512a-49c2-95f0-03ece023fe72")
     private val descriptiveXml = Paths.get(
         projectFolder,
-        "src/test/resources/tekst_ee11f8dd-512a-49c2-95f0-03ece023fe72-METS_BROWSING.xml"
+        "src/test/resources/tekst_ee11f8dd-512a-49c2-95f0-03ece023fe72/tekst_ee11f8dd-512a-49c2-95f0-03ece023fe72-METS_BROWSING.xml"
     )
 
     @BeforeEach
